fix sdk export of new method, put createBigNumber back in UI
