+++ packages/augur-ui/src/modules/forking/actions/get-fork-migration-totals.ts
@@ -1,5 +1,4 @@
 import logError from "utils/log-error";
-import { calculatePayoutNumeratorsValue } from "@augurproject/sdk/src/utils";
 import { AppState } from "store";
 import { NodeStyleCallback } from "modules/types";
 import { ThunkDispatch } from "redux-thunk";
