+++ packages/augur-ui/src/modules/markets-list/components/markets-view.tsx
@@ -11,6 +11,7 @@ interface MarketsViewProps {
   markets: Array<MarketData>;
   location: object;
   history: object;
+  isConnected: boolean,
   toggleFavorite: (...args: any[]) => any;
   loadMarketsInfoIfNotLoaded: (...args: any[]) => any;
   isMobile: boolean;
@@ -64,16 +65,16 @@ export default class MarketsView extends Component<
   }
 
   componentDidMount() {
-    const { universe } = this.props;
-    if (universe) {
+    const { isConnected } = this.props;
+    if (isConnected) {
       this.updateFilteredMarkets();
     }
   }
 
   componentDidUpdate(prevProps) {
-    const { search, category, universe } = this.props;
+    const { search, category, isConnected } = this.props;
     if (
-      universe !== prevProps.universe ||
+      isConnected !== prevProps.isConnected ||
       (search !== prevProps.search || category !== prevProps.category)
     ) {
       this.updateFilteredMarkets();
