+++ packages/augur-ui/src/modules/markets-list/containers/markets-view-container.ts
@@ -25,6 +25,7 @@ const mapStateToProps = (state: AppState, { location }) => {
   const searchPhrase = buildSearchString(keywords, selectedTagNames);
 
   return {
+    isConnected: state.connection.isConnected && state.universe.id != null,
     isLogged: state.authStatus.isLogged,
     universe: (state.universe || {}).id,
     search: searchPhrase,
