+++ packages/augur-ui/src/modules/reports/selectors/select-dispute-outcomes.ts
@@ -2,8 +2,8 @@ import {
   SCALAR,
   MALFORMED_OUTCOME,
   ZERO
-} from "modules/common/constants";import calculatePayoutNumeratorsValue from "utils/calculate-payout-numerators-value";
-import { calculatePayoutNumeratorsValue, createBigNumber } from "@augurproject/sdk/src/utils";
+} from "modules/common/constants";
+import { calculatePayoutNumeratorsValue } from "@augurproject/sdk/build";
 import { isEmpty } from "utils/is-populated";
 import { createBigNumber } from "utils/create-big-number";
