+++ packages/augur-ui/src/utils/convert-marketInfo-marketData.ts
@@ -11,8 +11,9 @@ import {
   CATEGORICAL,
 } from 'modules/common/constants';
 import { convertUnixToFormattedDate } from './format-date';
-import { calculatePayoutNumeratorsValue, createBigNumber } from "@augurproject/sdk/src/utils";
+import { calculatePayoutNumeratorsValue } from "@augurproject/sdk/build";
 import { formatPercent, formatDai, formatNone, formatNumber } from './format-number';
+import { createBigNumber } from './create-big-number';
 
 export function convertMarketInfoToMarketData(marketInfo: MarketInfo) {
   const reportingFee = parseInt(marketInfo.reportingFeeRate || '0', 10);
